Continuando sistema de autorização

+++ backend/src/main/java/com/estacionaai/backend/vaga/VagaController.java
@@ -1,15 +1,19 @@
 package com.estacionaai.backend.vaga;
 
+import com.estacionaai.backend.security.JwtService;
 import com.estacionaai.backend.user.User;
 import com.estacionaai.backend.user.UserRepository;
+import com.estacionaai.backend.user.UserService;
 import com.estacionaai.backend.vaga.dto.VagaCreateDTO;
 import com.estacionaai.backend.vaga.dto.VagaResponseDTO;
-import com.estacionaai.backend.vaga.dto.VagaUpdateDTO;
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
 import java.util.*;
 
 @RestController
@@ -19,28 +23,46 @@ public class VagaController {
     private VagaRepository repository;
     @Autowired
     private UserRepository userRepository;
+    @Autowired
+    private JwtService jwtService;
+    @Autowired
+    private UserService userService;
 
     @GetMapping("getAll")
     public List<Vaga> getAll() {
         return repository.findAll();
     }
 
-    @GetMapping("getById")
-    public VagaResponseDTO getById(@RequestParam(name = "id", required = true) UUID id) {
+    @GetMapping("{id}")
+    public VagaResponseDTO getById(@PathVariable UUID id) {
         return new VagaResponseDTO(repository.getReferenceById(id));
     }
 
-    @PostMapping("create")
-    public void create(@RequestBody VagaCreateDTO data){
+    @PostMapping
+    public HttpStatus create(HttpServletRequest request, @RequestBody VagaCreateDTO data){
+        String authHeader = request.getHeader(HttpHeaders.AUTHORIZATION);
+        String token = authHeader.substring(7);
+
+        String username = jwtService.getSubject(token);
+        var user = userRepository.findByUsername(username).get();
+
         Vaga vaga = new Vaga(data);
-        User user = userRepository.getReferenceById(data.owner());
         vaga.setOwner(user);
+
         repository.save(vaga);
+
+        return HttpStatus.CREATED;
     }
 
-    @DeleteMapping("delete")
-    public void delete(@RequestParam(name = "id", required = true) UUID id) {
-        repository.deleteById(id);
+    @DeleteMapping("{id}")
+    public HttpStatus delete(HttpServletRequest request, @PathVariable UUID id) {
+        User user = userService.getUserByRequest(request);
+        if(user.getId().equals(id)) {
+            repository.deleteById(id);
+            return HttpStatus.OK;
+        } else {
+            return HttpStatus.UNAUTHORIZED;
+        }
     }
 
 //    @PutMapping("update")
